requirements.txt
Django==2.1.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==0.0.8
django-superform==0.3.1
Pillow==5.4.1
pytz==2018.9
